+++ src/CHANGES.md
@@ -1,5 +1,5 @@
 **Added:**
-* none
+* export dialog
 
 **Changed:**
 * none